feat(phase2c+3): SharedKit test stabilization + iOS parity complete
Phase 2c — SharedKit Test Stabilization:
- Fix AnyCodable.encode() for [String: Any] / [Any] (was encoding as null)
- Fix GatewayChannelActor session injection (sessionBox not passed to channel)
- Fix GatewayNodeSession.connect() to forward sessionBox parameter
- Isolate DeviceAuthStore tests with temp dir for COREBLOW_STATE_DIR
- Relocate files to match OC structure (DiagnosticsFileLog, NodeMode/)
- Result: swift build 0 errors, swift test 190/190 pass

Phase 3 — iOS App Parity:
- Create 12 missing iOS files (CoreBlowApp, RootCanvas, SettingsTab, etc.)
- Expand 32 files to >=80% OC parity
- Expand iOS tests and ShareExtension
- Generate Xcode project (project.yml + CoreBlow.xcodeproj)
- iOS: 256 files, 21,689 lines (106% of OC 20,489)
- xcodebuild BUILD SUCCEEDED
- Zero OpenClaw naming remnants

+++ apps/macos/Sources/CoreBlow/CanvasWindowController+Window.swift
@@ -83,8 +83,8 @@ extension CanvasWindowController {
         // - If agent provides width/height, override size.
         // - If agent provides only size, keep the remembered origin.
         if let placement = self.preferredPlacement {
-            if let x = placement.x { frame.origin.x = x }
-            if let y = placement.y { frame.origin.y = y }
+            if let x = placement.x { frame.origin.x = CGFloat(x) }
+            if let y = placement.y { frame.origin.y = CGFloat(y) }
             if let w = placement.width { frame.size.width = max(CanvasLayout.minPanelSize.width, CGFloat(w)) }
             if let h = placement.height { frame.size.height = max(CanvasLayout.minPanelSize.height, CGFloat(h)) }
         }
